테스트 추가, 로컬 캐싱 1차 작업

+++ Scoop/Classes/util/Filesable.swift
@@ -11,7 +11,7 @@ import UIKit
  참고
  - link : http://iosbrain.com/blog/2018/04/22/ios-file-management-with-filemanager-in-protocol-oriented-swift-4/
  
- 앱 샌드박스 구조(간략)
+ App SandBox Struct(간략)
     Documents
     Library
         Caches
@@ -60,7 +60,7 @@ public extension Filesable {
         -   Parameter directory : 정의된 타입
         - Parameter createFolder : 폴더를 생성하냐? 1뎁스만 생성하자
      */
-    func getBaseDownloadURL(for directory: DIRECTORIES, createFolder: String? = nil) -> URL? {
+    internal func getBaseDownloadURL(for directory: DIRECTORIES = .documents, createFolder: String = "scoop") -> URL? {
         //SendBox내 기본 디렉토리 URL리턴
         func getTypeURL(type: DIRECTORIES) -> URL? {
             switch type {
@@ -71,70 +71,95 @@ public extension Filesable {
         }
         //기본디렉토리에 생상할 폴더가 있는지 체크해서 생성 후 URL 반환
         guard var typeURL = getTypeURL(type: directory) else {  return nil  }
-        if let folder = createFolder {
-            typeURL = typeURL.appendingPathComponent(folder)
-            do {
-                if typeURL.type == .not {
-                    try FileManager.default.createDirectory(at: typeURL, withIntermediateDirectories: true, attributes: nil)
-                }
-            } catch let error {
-                print("Could not copy file to disk: \(error.localizedDescription)")
-                return nil
+        typeURL = typeURL.appendingPathComponent(createFolder)
+        do {
+            if typeURL.type == .not {
+                try FileManager.default.createDirectory(at: typeURL, withIntermediateDirectories: true, attributes: nil)
             }
+        } catch let error {
+            print("Could not copy file to disk: \(error.localizedDescription)")
+            return nil
         }
         return typeURL
     }
     
-    func writeURL(basePath: URL?, requestURL: URL?) -> URL? {
+    /**
+     지정된 경로에 다운로드한 파일의 last path component 로 파일이 저장될 경로와 파일명 확장자까지 만듬
+     requestURL에 파일명 또는 확장자가 포함되지 않았다면 추가해줘야함
+     - Parameter basePath : 베이스가 되는 경로명
+     - Parameter requestURL : 다운로드한 요청한 URL
+     */
+    internal func writeURL(basePath: URL?, requestURL: URL?) -> URL? {
         
         guard let hasBasePath = basePath, let hasRequestURL = requestURL else { return nil }
         return hasBasePath.appendingPathComponent(hasRequestURL.lastPathComponent)
     }
     
-    //파일 이동, 복사, 삭제
     /**
-    move메서드는 overwrite를 기본으로 한다
-     -  Parameter from : 이동할 파일의 위치
+     로컬에 파일이 존재하는지 체크함
+     - Parameter at : 파일이 있을것으로 추정되는 로컬 경로
+     */
+    internal func fileExists(at: URL) -> Bool {
+        
+        return FileManager.default.fileExists(atPath: at.path)
+    }
+    
+    /**
+    파일을 이동한다, 중복파일에 대한 처리는 overwrite를 기본으로 한다
+     -  Parameter at : 이동할 파일의 위치
      - Parameter to : 이동될 파일의 위치
      */
-    func move(at: URL, to: URL, completeHandler: ((Bool) -> Void)? = nil) {
+    internal func move(at: URL, to: URL, completeHandler: ((Bool, Error?) -> Void)? = nil) {
         
         let manager = FileManager.default
         do {
-            try manager.removeItem(at: to)
+            if self.fileExists(at: to) {
+                try manager.removeItem(at: to)
+            }
             try manager.moveItem(at: at, to: to)
-            completeHandler?(true)
+            completeHandler?(true, nil)
             print("move sucess - \(to)")
         } catch let error {
             print("move error - \(error.localizedDescription)")
-            completeHandler?(false)
+            completeHandler?(false, error)
         }
     }
     
-    func copy(at: URL, to: URL, completeHandler: ((Bool) -> Void)? = nil) {
+    /**
+    파일을 복사한다, 중복파일에 대한 처리는 overwrite를 기본으로 한다
+     -  Parameter at : 복사할 파일의 위치
+     - Parameter to : 복사될 파일의 위치
+     */
+    internal func copy(at: URL, to: URL, completeHandler: ((Bool, Error?) -> Void)? = nil) {
         
         let manager = FileManager.default
         do {
-            try manager.removeItem(at: to)
+            if fileExists(at: to) {
+                try manager.removeItem(at: to)
+            }
             try manager.copyItem(at: at, to: to)
-            completeHandler?(true)
+            completeHandler?(true, nil)
             print("copy sucess - \(to)")
         } catch let error {
             print("copy error - \(error.localizedDescription)")
-            completeHandler?(false)
+            completeHandler?(false, error)
         }
     }
     
-    func remove(at: URL, completeHandler: ((Bool) -> Void)? = nil) {
+    /**
+    파일을 삭제한다,
+     -  Parameter at : 삭제할 파일의 위치
+     */
+    internal func remove(at: URL, completeHandler: ((Bool, Error?) -> Void)? = nil) {
         
         let manager = FileManager.default
         do {
             try manager.removeItem(at: at)
-            completeHandler?(true)
+            completeHandler?(true, nil)
             print("remove sucess - \(at)")
         } catch let error {
             print("remove error - \(error.localizedDescription)")
-            completeHandler?(false)
+            completeHandler?(false, error)
         }
     }
 }
